refactor(app, routes): simplify app setup / routing structure
