+++ packages/backend/src/App.js
@@ -1,45 +1,30 @@
 import express from "express";
 import cookieParser from "cookie-parser";
-import logger from "morgan";
+import morgan from "morgan";
 import cors from "cors";
 import { errorHandler } from "./apis/_helpers";
 import customJwt from "middlewares/jwt";
 import sessionRefresh from "middlewares/session-refresh";
-import { configureRoutes } from "_routes";
+import { configureRoutes } from "routes";
 
-/**
- * @param {object} expressApp Express application object.
- */
-export function configureExpress(expressApp) {
-  expressApp.use(express.json());
-  expressApp.use(
-    express.urlencoded({
-      extended: false,
-    })
-  );
-  expressApp.use(cookieParser());
-  expressApp.use(logger("dev"));
-  expressApp.use(
-    cors({
-      exposedHeaders: ["Authorization"],
-    })
-  );
-
-  expressApp.use(customJwt());
-  expressApp.use(sessionRefresh);
-}
-
-/**
- * @param {object} expressApp Express application object.
- */
-export function handleErrors(expressApp) {
-  expressApp.use(errorHandler);
-}
-
-// Initialize express app
 const app = express();
 
-configureExpress(app);
+app.use(express.json());
+app.use(
+  express.urlencoded({
+    extended: false,
+  })
+);
+app.use(cookieParser());
+app.use(morgan("dev"));
+app.use(errorHandler);
+app.use(
+  cors({
+    exposedHeaders: ["Authorization"],
+  })
+);
+app.use(customJwt());
+app.use(sessionRefresh);
 configureRoutes(app);
 
 export { app };
